CVP_MVSNet/requirements.txt
cycler==0.10.0
kiwisolver==1.1.0
matplotlib==3.1.3
numpy==1.18.1
opencv-python==4.2.0.32
Pillow==7.0.0
pkg-resources==0.0.0
plyfile==0.7.1
pyparsing==2.4.6
python-dateutil==2.8.1
six==1.14.0
torch==1.4.0
torchvision==0.5.0
